Align event counter factory with event domain
